Fixed: hash visit non-exists Int32 index raise error because hash key type is Int64

+++ src/i18n/zh_cn.cr
@@ -3,7 +3,7 @@ module TimeInWords
     module ZH_CN
       extend self
 
-      Map = Hash(Int64, String).new { |hash, k| hash[k] = "#{k} " }
+      Map = Hash(Int32, String).new { |hash, k| hash[k] = "#{k} " }
       Map[0] = "零"
       Map[1] = "一"
       Map[2] = "两"
